Extract reusable meeting detail pane
